+++ apps/docs/src/lib/components/blocks/app-sidebar/app-sidebar.svelte
