+++ apps/docs/src/lib/components/blocks/app-sidebar/nav-bottom.svelte
