+++ apps/docs/src/lib/components/blocks/app-sidebar/version-switcher.svelte
