+++ apps/docs/src/lib/components/blocks/toggle-scheme/toggle-scheme.svelte
